fix: closed position issue & code refactoring
