+++ src/components/ConcentratedLiquidity/PositionList/index.tsx
@@ -3,9 +3,11 @@ import Scrollbars from 'react-custom-scrollbars';
 import { Inbox, Search } from 'react-feather';
 import { useTranslation } from 'react-i18next';
 import { ThemeContext } from 'styled-components';
-import { Box, DropdownMenu, Loader, Text, TextInput } from 'src/components';
+import { Box, Button, DropdownMenu, Loader, Text, TextInput } from 'src/components';
+import { usePangolinWeb3 } from 'src/hooks';
+import { useWalletModalToggle } from 'src/state/papplication/hooks';
 import { Hidden } from 'src/theme/components';
-import { LoaderWrapper, MobileGridContainer, PanelWrapper, PoolsWrapper } from './styles';
+import { ErrorContainer, LoaderWrapper, MobileGridContainer, PanelWrapper, PoolsWrapper } from './styles';
 import { PositionListProps, SortingType } from './types';
 
 const PositionList: React.FC<PositionListProps> = (props) => {
@@ -15,6 +17,8 @@ const PositionList: React.FC<PositionListProps> = (props) => {
     label: t(`concentratedLiquidity.sortTypes.${SortingType[key]}`),
     value: SortingType[key],
   }));
+  const { account } = usePangolinWeb3();
+  const toggleWalletModal = useWalletModalToggle();
 
   const {
     isLoading,
@@ -86,12 +90,26 @@ const PositionList: React.FC<PositionListProps> = (props) => {
           </MobileGridContainer>
         </Box>
         {doesNotPoolExist ? (
-          <Box display={'flex'} flexDirection={'column'} alignItems={'center'}>
-            <Inbox size={'121px'} />
-            <Text pt={'25px'} pb={'25px'} color="color11" fontSize={[18, 22]} fontWeight={400}>
-              {t('concentratedLiquidity.positionNotFound')}
-            </Text>
-          </Box>
+          <ErrorContainer>
+            {account ? (
+              <Box display={'flex'} flexDirection={'column'} alignItems={'center'}>
+                <Inbox size={'121px'} />
+                <Text pt={'25px'} pb={'25px'} color="color11" fontSize={[18, 22]} fontWeight={400}>
+                  {t('concentratedLiquidity.positionNotFound')}
+                </Text>
+              </Box>
+            ) : (
+              <Box display={'flex'} flexDirection={'column'} alignItems={'center'}>
+                <Inbox size={'121px'} />
+                <Text pt={'25px'} pb={'25px'} color="color11" fontSize={[22, 26]} fontWeight={400}>
+                  {t('concentratedLiquidity.positionNotExist')}
+                </Text>
+                <Button width={'300px'} variant="primary" onClick={toggleWalletModal}>
+                  {t('common.connectWallet')}
+                </Button>
+              </Box>
+            )}
+          </ErrorContainer>
         ) : (
           <Scrollbars>
             <PanelWrapper>{children}</PanelWrapper>
